Moved self hosted GF outside of beacon-script file, to its own self-hosted-gf.ts file
